Added Api for refreshing token
- api/v1/auth/refresh-token
