+++ app/db/mysql.py
@@ -31,7 +31,7 @@ except OperationalError as e:
 
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/swagger")
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/auth/swagger")
 
 
 def get_db():
