+++ app/main.py
@@ -4,7 +4,7 @@
 from fastapi.staticfiles import StaticFiles
 
 from app.api.v1.endpoints import (
-    login_router,
+    auth_router,
     notes_router,
     users_router,
     home_router,
@@ -17,7 +17,7 @@ app.mount("/static", StaticFiles(directory="static"), name="static")
 
 app.include_router(home_router, tags=["home"])
 app.include_router(
-    login_router, prefix="/api/v1", tags=["Login"])
+    auth_router, prefix="/api/v1", tags=["Auth"])
 app.include_router(
     oauth_router, prefix="/api/v1", tags=["OAuth"])
 app.include_router(
